+++ lib/src/features/authentication/presentation/login/forgot_password_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
-import 'package:todo_app/src/core/appbar/custom_appbar.dart';
-import 'package:todo_app/src/core/extensions/snack_bar.dart';
-import 'package:todo_app/src/core/theme/todo_theme.dart';
+import 'package:todo_app/src/common_widgets/appbar/custom_appbar.dart';
+import 'package:todo_app/src/extensions/snack_bar.dart';
+import 'package:todo_app/src/constants/todo_theme.dart';
 import 'package:todo_app/src/features/authentication/presentation/widgets/auth_button.dart';
 import 'package:todo_app/src/features/authentication/data/auth_actions_provider.dart';
 import 'package:todo_app/src/features/authentication/data/auth_provider.dart';
