+++ lib/src/features/home/home_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:todo_app/src/core/bottom_navigation/bottom_navigation_provider.dart';
-import 'package:todo_app/src/core/bottom_navigation/bottom_navigation.dart';
+import 'package:todo_app/src/common_widgets/bottom_navigation/bottom_navigation_provider.dart';
+import 'package:todo_app/src/common_widgets/bottom_navigation/bottom_navigation.dart';
 import 'package:todo_app/src/features/completed/presentation/completed_page.dart';
 import 'package:todo_app/src/features/profile/presentation/profile_page.dart';
 import 'package:todo_app/src/features/todo/todo/todo_page.dart';
