+++ lib/src/features/todo/todo/todo_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:todo_app/src/core/appbar/custom_appbar.dart';
+import 'package:todo_app/src/common_widgets/appbar/custom_appbar.dart';
 import 'package:todo_app/src/common_widgets/error_page.dart';
 import 'package:todo_app/src/common_widgets/loading_page.dart';
 import 'package:todo_app/src/features/todo/todo/widgets/todo_card.dart';
